+++ packages/contracts/hardhat.config.ts
@@ -45,7 +45,6 @@ extendEnvironment(hre => {
   const aragonPluginRepos: AragonPluginRepos = {
     AddresslistVotingRepoProxy: '',
     TokenVotingRepoProxy: '',
-    MultisigRepoProxy: '',
   };
   const testingFork: TestingFork = {
     network: '',
@@ -54,7 +53,7 @@ extendEnvironment(hre => {
   };
   hre.aragonPluginRepos = aragonPluginRepos;
   hre.aragonToVerifyContracts = [];
-  hre.managementDAOMultisigPluginAddress = '';
+  hre.managementDAOMultisigPluginAddress = ''; // TODO
   hre.managementDAOActions = [];
   hre.testingFork = testingFork;
 });