+++ packages/contracts/test/deploy/managing-dao.ts
@@ -55,7 +55,6 @@ describe('Management DAO', function () {
   let pluginsRepos: {
     tokenVoting: PluginRepo;
     addresslistVoting: PluginRepo;
-    multisig: PluginRepo;
   };
 
   before(async () => {
@@ -140,10 +139,6 @@ describe('Management DAO', function () {
       ),
       addresslistVoting: PluginRepo__factory.connect(
         hre.aragonPluginRepos.AddresslistVotingRepoProxy,
-        deployer
-      ),
-      multisig: PluginRepo__factory.connect(
-        hre.aragonPluginRepos.MultisigRepoProxy,
         deployer
       ),
     };
@@ -291,30 +286,6 @@ describe('Management DAO', function () {
         expect(
           await pluginsRepos.addresslistVoting.isGranted(
             pluginsRepos.addresslistVoting.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-    });
-
-    context('Multisig Repo', async function () {
-      it('has permission to upgrade the repo', async function () {
-        expect(
-          await pluginsRepos.multisig.isGranted(
-            pluginsRepos.multisig.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.UPGRADE_REPO_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-
-      it('has permission to maintain the repo', async function () {
-        expect(
-          await pluginsRepos.multisig.isGranted(
-            pluginsRepos.multisig.address,
             managementDao.address,
             PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
             []